Potentially fix errors

+++ lib/ImageDriver/src/Image.h
@@ -7,7 +7,7 @@
  */
 class Image {
 public:
-    using Element = unsigned;
+    using Element = uint8_t;
     constexpr Image(const std::span<const Element> data, size_t width)
         : mData(data), mWidth(width), mHeight(data.size() / width) {
     }
